fix improper hook use + move button to new file
